+++ puma/__init__.py
@@ -2,7 +2,7 @@
 
 # flake8: noqa
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 
 from puma.histogram import Histogram, HistogramPlot
 from puma.integrated_eff import IntegratedEfficiency, IntegratedEfficiencyPlot