+++ changelog.md
@@ -2,6 +2,9 @@
 
 ### [Latest]
 
+- Add NaN filtering for HL plots. Bump atlas-ftag-tools version to 0.1.11. [!220](https://github.com/umami-hep/puma/pull/220)
+- Making Under- and Overflow bins default in histograms [!218](https://github.com/umami-hep/puma/pull/218)
+
 ### [v0.3.0] (2023/10/03)
 
 - Fix publish workflow, update tools package