+++ changelog.md
@@ -2,10 +2,14 @@
 
 ### [Latest]
 
+
+### [v0.3.1] (2023/11/27)
+
+- Update versions and add vscode settings [!226](https://github.com/umami-hep/puma/pull/226)
 - Update ratio uncertainty calculation [!225](https://github.com/umami-hep/puma/pull/225)
+- Adding option for initialising histograms with pre-binned distributions. [!221](https://github.com/umami-hep/puma/pull/221)
 - Add NaN filtering for HL plots. Bump atlas-ftag-tools version to 0.1.11. [!220](https://github.com/umami-hep/puma/pull/220)
 - Making Under- and Overflow bins default in histograms [!218](https://github.com/umami-hep/puma/pull/218)
-- Adding "bin_edges" and "sum_squared_weights" parameters to the Histogram constructor to define pre-filled histograms.
 
 ### [v0.3.0] (2023/10/03)